# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: RestMail API
  version: 1.0.0
  description: >
    RestMail is a modern email platform with REST API access, supporting traditional
    SMTP/IMAP/POP3 gateways and the RESTMAIL server-to-server protocol. This spec
    covers all user-facing, admin, internal gateway, and RESTMAIL endpoints.
  license:
    name: MIT
servers:
  - url: /
    description: Default server
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/accounts/{id}/quarantine:
    parameters:
      - $ref: "#/components/parameters/AccountId"
    get:
      operationId: listQuarantine
      tags: [Quarantine]
      summary: List quarantined messages
      responses:
        "200":
          description: Quarantined messages
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Quarantine"
  /api/v1/accounts/{id}/quarantine/{mid}/release:
    parameters:
      - $ref: "#/components/parameters/AccountId"
      - name: mid
        in: path
        required: true
        schema:
          type: integer
        description: Quarantine entry ID
    post:
      operationId: releaseQuarantine
      tags: [Quarantine]
      summary: Release a quarantined message to inbox
      responses:
        "200":
          description: Message released
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [status, message_id]
                    properties:
                      status:
                        type: string
                        example: released
                      message_id:
                        type: integer
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/accounts/{id}/quarantine/{mid}:
    parameters:
      - $ref: "#/components/parameters/AccountId"
      - name: mid
        in: path
        required: true
        schema:
          type: integer
        description: Quarantine entry ID
    delete:
      operationId: deleteQuarantine
      tags: [Quarantine]
      summary: Permanently delete a quarantined message
      responses:
        "200":
          description: Message deleted
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [status]
                    properties:
                      status:
                        type: string
                        example: deleted

  # ═══════════════════════════════════════════════════════════════
  # SSE
  # ═══════════════════════════════════════════════════════════════
...
  /api/v1/admin/pipelines:
    get:
      operationId: listPipelines
      tags: ["Admin: Pipelines"]
      summary: List pipelines, optionally filtered by domain
      parameters:
        - name: domain_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Pipeline list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Pipeline"
    post:
      operationId: createPipeline
      tags: ["Admin: Pipelines"]
      summary: Create a new pipeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePipelineRequest"
      responses:
        "201":
          description: Pipeline created
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Pipeline"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/admin/pipelines/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    patch:
      operationId: updatePipeline
      tags: ["Admin: Pipelines"]
      summary: Update a pipeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePipelineRequest"
      responses:
        "200":
          description: Pipeline updated
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Pipeline"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deletePipeline
      tags: ["Admin: Pipelines"]
      summary: Delete a pipeline
      responses:
        "200":
          description: Pipeline deleted
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [status]
                    properties:
                      status:
                        type: string
                        example: deleted
  /api/v1/admin/pipelines/test:
    post:
      operationId: testPipeline
      tags: ["Admin: Pipelines"]
      summary: Test a pipeline against a sample email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestPipelineRequest"
      responses:
        "200":
          description: Pipeline execution result
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/ExecutionResult"
  /api/v1/admin/pipelines/test-filter:
    post:
      operationId: testFilter
      tags: ["Admin: Pipelines"]
      summary: Test a single filter against a sample email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestFilterRequest"
      responses:
        "200":
          description: Filter execution result
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/FilterResult"
  /api/v1/admin/pipelines/logs:
    get:
      operationId: listPipelineLogs
      tags: ["Admin: Pipelines"]
      summary: List pipeline execution logs
      parameters:
        - name: pipeline_id
          in: query
          schema:
            type: integer
        - name: direction
          in: query
          schema:
            type: string
            enum: [inbound, outbound]
        - name: action
          in: query
          schema:
            type: string
            enum: [continue, reject, quarantine, discard]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Pipeline logs
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/PipelineLog"
                  pagination:
                    $ref: "#/components/schemas/Pagination"

  # ═══════════════════════════════════════════════════════════════
  # Admin: Custom Filters
  # ═══════════════════════════════════════════════════════════════
  /api/v1/admin/custom-filters:
    get:
      operationId: listCustomFilters
      tags: ["Admin: Custom Filters"]
      summary: List custom filters, optionally filtered by domain
      parameters:
        - name: domain_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Custom filter list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CustomFilter"
    post:
      operationId: createCustomFilter
      tags: ["Admin: Custom Filters"]
      summary: Create a custom filter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomFilterRequest"
      responses:
        "201":
          description: Custom filter created
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/CustomFilter"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/admin/custom-filters/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    delete:
      operationId: deleteCustomFilter
      tags: ["Admin: Custom Filters"]
      summary: Delete a custom filter
      responses:
        "200":
          description: Custom filter deleted
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [status]
                    properties:
                      status:
                        type: string
                        example: deleted

  # ═══════════════════════════════════════════════════════════════
  # Admin: Queue
  # ═══════════════════════════════════════════════════════════════
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/v1/auth/login
  parameters:
    AccountId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Account ID
...
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Resource ID
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: not_found
            message:
              type: string
              example: Resource not found
            details:
              type: object
    ValidationErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: validation_failed
            message:
              type: string
              example: Request body failed validation
            details:
              type: object
              properties:
                fields:
                  type: object
                  additionalProperties:
                    type: string
    Pagination:
      type: object
      properties:
        cursor:
          type: string
          description: Opaque cursor for the next page
        has_more:
          type: boolean
        total:
          type: integer
          format: int64

    # ─────────────────────────────────────────────────────────────
    # Health
    # ─────────────────────────────────────────────────────────────
...
    Quarantine:
      type: object
      required: [id, mailbox_id, sender, subject, body_preview, quarantine_reason, received_at, expires_at, released]
      properties:
        id:
          type: integer
        mailbox_id:
          type: integer
        sender:
          type: string
        subject:
          type: string
        body_preview:
          type: string
        spam_score:
          type: number
          format: float
          nullable: true
        quarantine_reason:
          type: string
        received_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        released:
          type: boolean

    # ─────────────────────────────────────────────────────────────
    # Inbound / Gateway
    # ─────────────────────────────────────────────────────────────
...
    Pipeline:
      type: object
      required: [id, domain_id, direction, filters, active, created_at, updated_at]
      properties:
        id:
          type: integer
        domain_id:
          type: integer
        direction:
          type: string
          enum: [inbound, outbound]
        filters:
          description: Ordered array of filter configurations
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              config:
                type: object
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreatePipelineRequest:
      type: object
      required: [domain_id, direction, filters]
      properties:
        domain_id:
          type: integer
        direction:
          type: string
          enum: [inbound, outbound]
        filters:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              config:
                type: object
        active:
          type: boolean
          default: true
    UpdatePipelineRequest:
      type: object
      properties:
        filters:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              config:
                type: object
        active:
          type: boolean
    PipelineLog:
      type: object
      required: [id, pipeline_id, direction, action, steps, duration_ms, created_at]
      properties:
        id:
          type: integer
        pipeline_id:
          type: integer
        message_id:
          type: integer
          nullable: true
        direction:
          type: string
          enum: [inbound, outbound]
        action:
          type: string
          enum: [continue, reject, quarantine, discard]
        steps:
          type: array
          items:
            type: object
            description: Individual filter execution step
        duration_ms:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
    TestPipelineRequest:
      type: object
      required: [pipeline_id, email]
      properties:
        pipeline_id:
          type: integer
        email:
          type: object
          description: Sample email JSON for testing
    TestFilterRequest:
      type: object
      required: [filter_name, config, email]
      properties:
        filter_name:
          type: string
        config:
          type: object
        email:
          type: object
          description: Sample email JSON for testing
    ExecutionResult:
      type: object
      required: [final_action, steps, duration_ms]
      properties:
        final_action:
          type: string
          enum: [continue, reject, quarantine, discard]
        steps:
          type: array
          items:
            type: object
        duration_ms:
          type: integer
          format: int64
        final_email:
          type: object
          description: The email after all transformations
    FilterResult:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [continue, reject, quarantine, discard]
        email:
          type: object
          description: The email after filter transformation
        metadata:
          type: object

    # ─────────────────────────────────────────────────────────────
    # Custom Filter
    # ─────────────────────────────────────────────────────────────
    CustomFilter:
      type: object
      required: [id, domain_id, name, filter_type, direction, config, enabled, created_at, updated_at]
      properties:
        id:
          type: integer
        domain_id:
          type: integer
        name:
          type: string
        description:
          type: string
        filter_type:
          type: string
          enum: [action, transform]
        direction:
          type: string
          enum: [inbound, outbound, both]
        config:
          type: object
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateCustomFilterRequest:
      type: object
      required: [domain_id, name, filter_type, direction, config]
      properties:
        domain_id:
          type: integer
        name:
          type: string
        description:
          type: string
        filter_type:
          type: string
          enum: [action, transform]
        direction:
          type: string
          enum: [inbound, outbound, both]
        config:
          type: object
        enabled:
          type: boolean
          default: true

    # ─────────────────────────────────────────────────────────────
    # Outbound Queue
    # ─────────────────────────────────────────────────────────────
...